python/requirements.txt
torch
numpy
facenet_pytorch
Pillow
opencv-python
cmake
face-recognition